// src/auth/provider/services/lib/interfaces/google-profile.interface.ts
export interface GoogleProfile extends Record<string, any> {
    aud: string;
    azp: string;
    email: string;
    email_verified: boolean;
    exp: number;
    family_name?: string;
    given_name: string;
    hd?: string;
    iat: number;
    iss: string;
    jti?: string;
    name: string;
    nbf?: number;
    picture: string;
    sub: string;
    access_token: string;
    refresh_token?: string

}